+++ recipes/default.rb
@@ -1,5 +1,5 @@
 #
-# Cookbook Name:: ulimit
+# Cookbook Name:: ulimit2
 # Recipe:: default
 #
 # Copyright 2014, Mike Morris
@@ -7,7 +7,7 @@
 
 template ::File.join(node['ulimit']['conf_dir'], node['ulimit']['conf_file']) do
   source 'limits.conf.erb'
-  cookbook 'ulimit'
+  cookbook 'ulimit2'
   owner 'root'
   group 'root'
   mode  '0644'