+++ CHANGELOG.md
@@ -1,9 +1,9 @@
-ulimit CHANGELOG
+ulimit2 CHANGELOG
 ================
 
-This file is used to list changes made in each version of the ulimit cookbook.
+This file is used to list changes made in each version of the ulimit2 cookbook.
 
 0.1.0
 -----
-- [michael.m.morris@gmail.com] - Initial release of ulimit cookbook
+- [michael.m.morris@gmail.com] - Initial release of ulimit2 cookbook